Create a new frontend eslint config without warnings.
Turn on autofixable and easy fixable warnings, turn off the rest.
This is the first batch of the rules.
As checked with a symbolic link in frontier, most of the warnings are autofixable.
A small set, in frontier it's around 130, should be fixed manually. However, these fixes will be really easy.

- added webstorm working folder to gitignore

BPA-533